requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
et-xmlfile==1.0.1
jdcal==1.3
numpy==1.13.0rc1+mkl
olefile==0.44
openpyxl==2.4.7
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytesseract==0.1.6
scipy==0.19.0
selenium==3.4.1
six==1.10.0
